+++ CHANGELOG.md
@@ -3,9 +3,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 7.2.2 - 2021-05-27
+### Fixed
+- `hive-hooks` component to convert Storage Descriptor Info 'path' parameter (instead of Storage Descriptor).
+
 ## 7.2.1 - 2021-05-25
 ### Changed
-- `hive-hooks` component to convert Storage Description 'path' parameter.
+- `hive-hooks` component to convert Storage Descriptor 'path' parameter.
 - `hive-hooks` component to convert table parameter 'avro.schema.url'.
 
 ### Removed